added Webdriver manager class and modified all tests
