+++ src/main/java/HotelBookingTest.java
@@ -1,17 +1,20 @@
-import com.sun.javafx.PlatformUtil;
 import java.util.List;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.Select;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 public class HotelBookingTest {
 
     WebDriver driver;
-
+    WebDriverManager webDriverManager;
+    
+    
+    
     @FindBy(xpath = "//*[@id='Home']/div/aside[1]/nav/ul[1]/li[2]/a[1]")
     private WebElement hotelLink;
 
@@ -27,11 +30,17 @@ public class HotelBookingTest {
     @FindBy(xpath = "//*[@id='ui-id-1']//*[@class='list']//child::a")
     private List<WebElement> locations;
   
+    
+    @BeforeTest
+	public void setup()
+	{
+		webDriverManager = new WebDriverManager();
+		driver = webDriverManager.getDriver();
+	}
 
     @Test
     public void shouldBeAbleToSearchForHotels() {
-        setDriverPath();
-        driver = new ChromeDriver();
+        
         PageFactory.initElements(driver, this);
         driver.get("https://www.cleartrip.com/");
         driver.manage().window().maximize();
@@ -42,21 +51,16 @@ public class HotelBookingTest {
         locations.get(0).click();
         new Select(travellerSelection).selectByVisibleText("1 room, 2 adults");        
         searchButton.click();
-        driver.quit();
+        
 
     }
 
-    private void setDriverPath() {
-        if (PlatformUtil.isMac()) {
-            System.setProperty("webdriver.chrome.driver", "chromedriver");
-        }
-        if (PlatformUtil.isWindows()) {
-            System.setProperty("webdriver.chrome.driver", "chromedriver.exe");
-        }
-        if (PlatformUtil.isLinux()) {
-            System.setProperty("webdriver.chrome.driver", "chromedriver_linux");
-        }
-    }
+    @AfterTest
+   	public void tearDown()
+   	{
+   		
+       	webDriverManager.closeDriver();
+   	}
     
     private void waitFor(int durationInMilliSeconds) {
         try {
