+++ src/main/java/SignInTest.java
@@ -1,10 +1,8 @@
-import com.sun.javafx.PlatformUtil;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 
@@ -12,15 +10,20 @@ public class SignInTest {
 
 	    
 	WebDriver driver;
+	WebDriverManager webDriverManager;
+	
+	@BeforeTest
+	public void setup()
+	{
+		webDriverManager = new WebDriverManager();
+		driver = webDriverManager.getDriver();
+	}
 
     @Test
     public void shouldThrowAnErrorIfSignInDetailsAreMissing() {
-
-    	setDriverPath();        
-        driver = new ChromeDriver();
-        driver.get("https://www.cleartrip.com/");
+    	
+    	driver.get("https://www.cleartrip.com/");
         waitFor(2000);
-        driver.manage().window().maximize();
         driver.findElement(By.linkText("Your trips")).click();
         waitFor(2000);
         driver.findElement(By.id("SignIn")).click();
@@ -32,8 +35,15 @@ public class SignInTest {
         waitFor(2000);
         String errors1 = driver.findElement(By.id("errors1")).getText();
         Assert.assertTrue(errors1.contains("There were errors in your submission"));
-        driver.quit();
+        
     }
+    
+    @AfterTest
+	public void tearDown()
+	{
+		
+    	webDriverManager.closeDriver();
+	}
 
     private void waitFor(int durationInMilliSeconds) {
         try {
@@ -43,17 +53,7 @@ public class SignInTest {
         }
     }
 
-    private void setDriverPath() {
-        if (PlatformUtil.isMac()) {
-            System.setProperty("webdriver.chrome.driver", "chromedriver");
-        }
-        if (PlatformUtil.isWindows()) {
-            System.setProperty("webdriver.chrome.driver", "chromedriver.exe");
-        }
-        if (PlatformUtil.isLinux()) {
-            System.setProperty("webdriver.chrome.driver", "chromedriver_linux");
-        }
-    }
+    
 
 
 }
